+++ CHANGES.md
@@ -1,6 +1,9 @@
 [unreleased]
 * remove constructor
 * add some tests
+* add filter `gu_get_repo_api`
+* add filter `gu_get_auth_header`
+* add filter `gu_post_get_credentials`
 
 #### 0.5.0 / 2021-03-12
 * de-anonymize hooks